fix issues with FinishImportSection
FinishImportSection() had an issue with locking that effectively made it single-threaded because of how long inserting takes. This has been fixed by using a bulk insert :: Upsert(IEnumerable<t>. This is something that I had no idea existed as it is not documented anywhere to my knowledge, and I found it while looking through the source code.
I also changed the relevant functions to only add the importId to a HashInfo once the import section has fully finished. This way if the program exits before the section finishes it will not count the imageHashes as 'ignored' (because they will not have the current importId already).
There could still be an issue with counts being incorrect, but I could not make this happen with tests. If it does happen then just re-importing the folder and deleting the tab from the first attempt would fix the issue. (Re-importing is very fast).
Also fixed a crash related to animated images.